Closes #849 

Using `GeometryMapping` in `BCValues` slightly reduces code duplication and makes the similarity (and differences) between `BCValues` and `FaceValues` clearer. Not sure why faster on construction, but for `update!` it probably comes from the performance annotations used following the `FaceValues` setup with inbounds and inlines. 

Updated benchmark after #860, values from run 2 below

**Construction**
* PR: 18.499 ms (4266 allocations: 93.64 MiB)
* master: 20.720 ms (4258 allocations: 93.64 MiB)

**`update!(ch, t)`**
* PR: 333.900 μs (18 allocations: 1.47 KiB)
* master: 373.200 μs (20 allocations: 1.53 KiB)


<details> <summary> Benchmark script and results </summary>

```julia
using Ferrite, BenchmarkTools


grid = generate_grid(Quadrilateral, (1000, 1000))
function make_dh(grid)    
    dh = DofHandler(grid)
    ip = Lagrange{RefQuadrilateral,2}()
    add!(dh, :u, ip^2)
    add!(dh, :v, ip)
    close!(dh)
    return dh
end
dh = make_dh(grid)

function make_ch(dh, grid)
    ch = ConstraintHandler(dh)
    add!(ch, Dirichlet(:u, getfaceset(grid, "right"), Returns(zero(Vec{2}))))
    add!(ch, Dirichlet(:v, getfaceset(grid, "left"), Returns(0.0)))
    close!(ch)
    return ch
end

ch = make_ch(dh, grid)
@btime make_ch($dh, $grid)

update!(ch, 1.0)

@btime update!($ch, 2.0)

#= 
>>> PR
run 1
  18.414 ms (4266 allocations: 93.64 MiB)
  341.900 μs (18 allocations: 1.47 KiB)

run 2
  18.499 ms (4266 allocations: 93.64 MiB)
  333.900 μs (18 allocations: 1.47 KiB)

>>> master
run 1
    22.356 ms (4258 allocations: 93.64 MiB)
  367.800 μs (20 allocations: 1.53 KiB)

run 2
  20.720 ms (4258 allocations: 93.64 MiB)
  373.200 μs (20 allocations: 1.53 KiB)

=#
```

</details>