@KnutAM 2023-12-17
> This breaks edge boundary conditions, right?

That should work fine because we use the appropriate "dof-locations" based on `dirichlet_boundarydof_indices(boundary_type)(func_interpol)`. I didn't check this specifically, but I assume these cases are covered by the test suite.