@KnutAM 2025-02-15
In view of #1152 I think this might be the wrong direction actually.